Fixed stuff

+++ src/hyper/Pool.java
@@ -413,7 +413,7 @@ public class Pool implements Serializable {
 
     public boolean alreadyMeasured() {
 
-        return (this.Species.get(this.currentSpecies - 1).Genomes.get(this.currentGenome - 1).getFitness() != 0);
+        return (this.Species.get(this.currentSpecies - 1).Genomes.get(this.currentGenome - 1).getFitness() != 0L);
     }
     
     public Genome currentGenome(){
@@ -432,12 +432,12 @@ public class Pool implements Serializable {
             out.writeInt(this.Outputs);
             out.writeInt(this.generation);
             out.writeInt(this.Innovation);
-            out.writeFloat(this.maxFitness);
+            out.writeInt(this.maxFitness);
             out.writeInt(this.Species.size());
             for (Species sp : this.Species) {
                 out.writeInt(sp.Genomes.size());
-                out.writeFloat(sp.getTopFitness());
-                out.writeFloat(sp.getAverageFitness());
+                out.writeLong(sp.getTopFitness());
+                out.writeLong(sp.getAverageFitness());
                 for (Genome g : sp.Genomes) {
                     out.writeObject(g);
                 }
@@ -492,8 +492,8 @@ public class Pool implements Serializable {
             for (int i = 0; i < Speciessize; i++) {
                 int Genomesize = in.readInt();
                 Species spe = new Species(p.Inputs, p.Outputs, p);
-                spe.setTopFitness(in.readInt());
-                spe.setAverageFitness(in.readInt());
+                spe.setTopFitness(in.readLong());
+                spe.setAverageFitness(in.readLong());
                 for (int j = 0; j < Genomesize; j++) {
                     Genome g = (Genome) in.readObject();
                     g.setParent(p);
